+++ lib/wot_td_widgets.dart
@@ -1,0 +1,15 @@
+// Copyright 2022 The NAMIB Project Developers. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+//
+// SPDX-License-Identifier: BSD-3-Clause
+
+/// Experimental Widgets for UI Elements generated from W3C Web of Things
+/// Thing Descriptions (WoT TD). Based on the WoT Scripting API implementation
+/// dart_wot.
+library wot_td_widgets;
+
+export 'package:dart_wot/dart_wot.dart';
+
+export 'src/thing_widget.dart';
+export 'src/thing_card.dart';