+++ test/wot_td_widgets_test.dart
@@ -1,0 +1,15 @@
+// Copyright 2022 The NAMIB Project Developers. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+//
+// SPDX-License-Identifier: BSD-3-Clause
+
+import 'package:flutter_test/flutter_test.dart';
+
+// import 'package:wot_td_widgets/wot_td_widgets.dart';
+
+void main() {
+  test('Empty Test', () {
+    // TODO(JKRhb): Add test
+  });
+}